feat(controller+command+service+mapper+xml): 版本提升为0.5.0，实现监听上传视频->自动发送弹幕功能，简单完成弹幕发送中断和恢复功能，简单完成手动添加弹幕视频功能，调整数据库结构
版本提升为0.5.0

实现监听上传视频->自动发送弹幕功能
简单完成弹幕发送中断和恢复功能
简单完成手动添加弹幕视频功能

调整数据库结构

+++ src/main/java/com/github/cuteluobo/livedanmuarchive/utils/BiliLoginUtil.java
@@ -76,7 +76,7 @@ public class BiliLoginUtil {
             JsonNode dataNode = bodyNode.get("data");
             //uid
             JsonNode midNode = dataNode.get("mid");
-            baseUserInfo.setUid(midNode.asInt());
+            baseUserInfo.setUid(midNode.asLong());
             //等级
             JsonNode levelNode = dataNode.get("level_info");
             JsonNode currentLevelNode = levelNode.get("current_level");
